Refactor: Move tool implementations to a separate directory
Moved the `get_weather` and `get_current_time` tool implementations from `agent.py` into a new `tools` directory. This improves the organization of the code by separating the agent definition from the tool implementations.
